requirements.txt
asgiref==3.7.2
Django==5.0.1
sqlparse==0.4.4
Pillow
djangorestframework==3.14.0
markdown==3.5
django-filter==23.2
drf-yasg==1.21.7